dalvik: Fix for broswer-crashing bug
This applies the fix proposed by Olivier Cozette in
https://code.google.com/p/android/issues/detail?id=37561
and fixes our bug:
https://bugs.launchpad.net/linaro-android/+bug/978060

+++ vm/compiler/codegen/arm/CodegenDriver.cpp
@@ -2841,7 +2841,7 @@ static s8 findPackedSwitchIndex(const u2* switchData, int testVal, int pc)
     }
 
     chainingPC += jumpIndex * CHAIN_CELL_NORMAL_SIZE;
-    return (((s8) caseDPCOffset) << 32) | (u8) chainingPC;
+    return (((s8) caseDPCOffset) << 32) | (u8) ((unsigned int)chainingPC);
 }
 
 /* See comments for findPackedSwitchIndex */
